federated-dr/requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Federated Learning
flwr==1.5.0

# Image Models
timm>=1.0.0

# Privacy
opacus>=1.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.21.0
pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Development (optional)
jupyter>=1.0.0
notebook>=7.0.0

